Refactor: User 도메인 실행되게 수정
