+++ moodiary/src/main/java/com/clover/moodiary/user/command/controller/UserCommandController.java
@@ -1,8 +1,6 @@
-package com.clover.moodiary.user.command.contorller;
+package com.clover.moodiary.user.command.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
