+++ moodiary/src/main/java/com/clover/moodiary/user/command/service/UserCommandService.java
@@ -1,4 +1,7 @@
 package com.clover.moodiary.user.command.service;
 
+import org.springframework.stereotype.Service;
+
+@Service
 public interface UserCommandService {
 }
