+++ moodiary/src/main/java/com/clover/moodiary/user/command/service/UserCommandServiceImpl.java
@@ -3,5 +3,5 @@ package com.clover.moodiary.user.command.service;
 import org.springframework.stereotype.Service;
 
 @Service
-public class UserCommandServiceImpl {
+public class UserCommandServiceImpl implements UserCommandService {
 }
